RA-1174 - :hammer: Moving GetTraineeship vacancy tests under its own feature unit tests folder.
